+++ packages/terra-form-select/src/Combobox.jsx
@@ -89,7 +89,7 @@ const propTypes = {
    */
   optionFilter: PropTypes.func,
   /**
-   * Placeholder text.
+   * Placeholder text. defaults to 'Select or Enter'
    */
   placeholder: PropTypes.string,
   /**
@@ -143,6 +143,7 @@ class Combobox extends React.Component {
     this.handleChange = this.handleChange.bind(this);
     this.handleDeselect = this.handleDeselect.bind(this);
     this.handleSelect = this.handleSelect.bind(this);
+    this.handleResetComboboxValue = this.handleResetComboboxValue.bind(this);
   }
 
   /**
@@ -190,6 +191,13 @@ class Combobox extends React.Component {
   }
 
   /**
+   * Added to reset value field for combobox
+   */
+  handleResetComboboxValue() {
+    this.setState({ value: '' });
+  }
+
+  /**
    * Returns the appropriate variant display
    */
   display() {
@@ -203,7 +211,7 @@ class Combobox extends React.Component {
       allowClear, children, defaultValue, onChange, placeholder, required, value, intl, inputId, ...otherProps
     } = this.props;
 
-    const defaultPlaceholder = intl.formatMessage({ id: 'Terra.form.select.defaultDisplay' });
+    const defaultPlaceholder = intl.formatMessage({ id: 'Terra.form.select.defaultComboboxDisplay' });
     const selectPlaceholder = placeholder === undefined ? defaultPlaceholder : placeholder;
     let clearOptionDisplay;
 
@@ -228,6 +236,7 @@ class Combobox extends React.Component {
         totalOptions={SelectUtil.getTotalNumberOfOptions(children)}
         clearOptionDisplay={clearOptionDisplay}
         inputId={inputId}
+        resetComboboxValue={this.handleResetComboboxValue}
       >
         {this.state.tags}
         {children}