+++ packages/terra-form-select/src/search/Frame.jsx
@@ -115,6 +115,11 @@ const propTypes = {
    * The select value.
    */
   value: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.array]),
+  /**
+   * @private
+   * Callback function to reset input value after search
+   */
+  resetComboboxValue: PropTypes.func,
 };
 
 const defaultProps = {
@@ -173,11 +178,11 @@ class Frame extends React.Component {
     this.handleToggleMouseDown = this.handleToggleMouseDown.bind(this);
     this.handleToggleButtonMouseDown = this.handleToggleButtonMouseDown.bind(this);
     this.handleTouchStart = this.handleTouchStart.bind(this);
-    this.role = this.role.bind(this);
     this.menuId = `terra-select-menu-${uniqueid()}`;
     this.visuallyHiddenComponent = React.createRef();
     this.setSelectMenuRef = this.setSelectMenuRef.bind(this);
     this.shouldFocusDropdown = false;
+    this.hasEscPressed = false;
   }
 
   componentDidMount() {
@@ -257,8 +262,19 @@ class Frame extends React.Component {
       event.preventDefault();
       this.openDropdown(event);
     } else if (this.state.isOpen && keyCode === KeyCode.KEY_ESCAPE) {
+      this.hasEscPressed = true;
       event.stopPropagation();
       this.closeDropdown();
+    } else if (!this.state.isOpen && keyCode === KeyCode.KEY_ESCAPE && this.hasEscPressed) {
+      this.hasEscPressed = false;
+      if (this.props.resetComboboxValue) {
+        this.props.resetComboboxValue();
+      }
+      this.setState({
+        hasSearchChanged: false,
+        searchValue: '',
+      });
+      event.stopPropagation();
     }
   }
 
@@ -371,7 +387,7 @@ class Frame extends React.Component {
   getDisplay(ariaDescribedBy, id) {
     const { hasSearchChanged, searchValue } = this.state;
     const {
-      disabled, display, placeholder, required, inputId,
+      disabled, display, placeholder, required, inputId, isInvalid,
     } = this.props;
 
     const inputAttrs = {
@@ -386,11 +402,14 @@ class Frame extends React.Component {
       'aria-describedby': ariaDescribedBy,
       'aria-disabled': disabled,
       'aria-owns': this.state.isOpen ? id : undefined,
+      'aria-expanded': !disabled && this.state.isOpen,
       type: 'text',
       className: cx('search-input', { 'is-hidden': FrameUtil.shouldHideSearch(this.props, this.state) }),
       required,
       'aria-required': required,
       id: inputId,
+      role: 'combobox',
+      'aria-invalid': isInvalid,
     };
     const value = hasSearchChanged ? searchValue : display;
 
@@ -410,8 +429,6 @@ class Frame extends React.Component {
       isFocused: document.activeElement === this.input || document.activeElement === this.select,
       isOpen: false,
       isPositioned: false,
-      hasSearchChanged: false,
-      searchValue: '',
     });
   }
 
@@ -505,15 +522,6 @@ class Frame extends React.Component {
     }
 
     return ariaLabel === undefined ? defaultAriaLabel : ariaLabel;
-  }
-
-  /**
-   * Determines compatible role attribute to apply to select based on active variant and disabled prop
-   */
-  role() {
-    const { disabled } = this.props;
-    // role="application" needed to allow JAWS to work correctly with the select and use our key event listeners
-    return disabled ? undefined : 'application';
   }
 
   /**
@@ -670,14 +678,10 @@ class Frame extends React.Component {
     return (
       <div
         {...customProps}
-        role={this.role()}
         data-terra-select-combobox
         aria-controls={!disabled && this.state.isOpen ? this.menuId : undefined}
         aria-disabled={!!disabled}
-        aria-expanded={!!disabled && !!this.state.isOpen}
         aria-haspopup={!disabled ? 'true' : undefined}
-        aria-describedby={ariaDescribedBy}
-        aria-owns={this.state.isOpen ? this.menuId : undefined}
         className={selectClasses}
         onBlur={this.handleBlur}
         onFocus={this.handleFocus}