+++ packages/terra-form-select/src/shared/_SearchResults.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import classNamesBind from 'classnames/bind';
+import ThemeContext from 'terra-theme-context';
+import styles from './_SearchResults.module.scss';
+
+const cx = classNamesBind.bind(styles);
+
+const propTypes = {
+  /**
+   * Content to display when search results are found.
+   */
+  searchResultContent: PropTypes.node,
+};
+
+const SearchResults = ({ searchResultContent }) => {
+  const theme = React.useContext(ThemeContext);
+  return (
+    <div role="option" className={cx('results-that-contain', theme.className)} aria-selected="false">
+      {searchResultContent}
+    </div>
+  );
+};
+
+SearchResults.propTypes = propTypes;
+
+export default SearchResults;