+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,7 +3,17 @@
 ## Unreleased
 
 * Added
+  * Added visual label and associated with input field using 'for' and 'id'.
+## 1.51.0 - (November 16, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.50.0 - (November 13, 2023)
+
+* Added
   * Added about page for `TagList`.
+  * Updated doc examples for `terra-badge`.
 
 ## 1.49.0 - (November 3, 2023)