+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 * Added
+  * Added 'aria-invalid' attribute which will be set to true for error input fields and false when resolving errors. 
+
+## 6.50.0 - (November 13, 2023)
+
+* Changed
+  * Changed role from application to combobox in `terra-form-select-combobox`.
+  * Updated keydown handler to clear the text on `ESC` key press in `terra-form-select-combobox`.
+
+* Added
+  * Added results that contain in `terra-form-select-combobox`.
   * Added No results found text in `terra-form-select-combobox`.
 
 ## 6.49.0 - (October 23, 2023)
@@ -14,7 +24,7 @@
 
 * Changed
   * Fixed Cyclic navigation and added aria-required to single select.
-  * Added relevant roles to fix a11y issues. 
+  * Added relevant roles to fix a11y issues.
 
 ## 6.47.0 - (September 21, 2023)