+++ packages/terra-hyperlink/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 2.65.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.64.0 - (October 16, 2023)
 
 * Changed